moved histogram bucket generators to stats_utils package
