+++ api/agent/drivers/stats/stats_test.go
@@ -5,7 +5,6 @@ import (
 	"time"
 
 	"github.com/fnproject/fn/api/common"
-	"github.com/stretchr/testify/assert"
 )
 
 func TestAverage(t *testing.T) {
@@ -92,16 +91,4 @@ func TestDecimate(t *testing.T) {
 	if len(stats) != 49 {
 		t.Error("decimate function bad", len(stats))
 	}
-}
-
-func TestGenerateLogScaleHistogramBucketsWithRange(t *testing.T) {
-	assert.Equal(t, []float64{7, 14, 28, 56, 112}, GenerateLogScaleHistogramBucketsWithRange(7, 100))
-}
-
-func TestGenerateLinearHistogramBuckets(t *testing.T) {
-	assert.Equal(t, []float64{5, 7, 9, 11, 13, 15}, GenerateLinearHistogramBuckets(5, 15, 5))
-}
-
-func TestGenerateLogScaleHistogramBuckets(t *testing.T) {
-	assert.Equal(t, []float64{0, 0.1953125, 0.390625, 0.78125, 1.5625, 3.125, 6.25, 12.5, 25, 50, 100}, GenerateLogScaleHistogramBuckets(100, 10))
 }
